requirements.txt
python-telegram-bot==21.9
requests>=2.31.0
